pyproject.toml
[project]
name = "church-budget-tool"
version = "0.1.0"
description = "Church Budget Tool — FastAPI + Jinja2 + htmx dashboard for church financial management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx",  # for TestClient
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
